delete LCAO_hamilt.hpp and add spar_exx.cpp
